Trait ncomm_core::publisher_subscriber::Publisher

pub trait Publisher {
    type Data;
    type Error;

    // Required method
    fn publish(&mut self, data: Self::Data) -> Result<(), Self::Error>;
}

The basic publisher trait that enables the publishing of data to some endpoint for subscribers to read.

Required Associated Types

type Data

The data to be published by the publisher

type Error

The error type from attempting to publish data

Required Methods

fn publish(&mut self, data: Self::Data) -> Result<(), Self::Error>

Publish a piece of data to the endpoint for clients to read.